refactor to group adapter into a folder
