+++ ui/desktop/src/acp/adapter/messages.ts
@@ -1,0 +1,162 @@
+import type {
+  ContentBlock as AcpContentBlock,
+  SessionNotification,
+} from '@agentclientprotocol/sdk';
+import type { Message, MessageContent } from '../../api';
+import {
+  type AcpChatStateChange,
+  type AdapterState,
+  DEFAULT_VISIBLE_MESSAGE_METADATA,
+  getGooseMessageMeta,
+  messagesChange,
+} from './shared';
+
+export function applyContentChunk(
+  state: AdapterState,
+  role: Message['role'],
+  update: Extract<
+    SessionNotification['update'],
+    { sessionUpdate: 'user_message_chunk' | 'agent_message_chunk' }
+  >
+): AcpChatStateChange[] {
+  const content = messageContentFromAcpContentBlock(update.content);
+  if (!content) {
+    return [];
+  }
+
+  const gooseMeta = getGooseMessageMeta(update);
+  const messageId = update.messageId ?? gooseMeta.messageId;
+  const existing = findMessageForChunk(state, role, messageId, gooseMeta.created);
+
+  if (existing) {
+    const lastContent = existing.content[existing.content.length - 1];
+    if (lastContent?.type === 'text' && content.type === 'text') {
+      lastContent.text = mergeTextChunk(lastContent.text, content.text);
+    } else if (content.type === 'image' && hasImageContent(existing, content)) {
+      return messagesChange(state);
+    } else {
+      existing.content.push(content);
+    }
+  } else {
+    state.messages.push({
+      ...(messageId ? { id: messageId } : {}),
+      role,
+      created: gooseMeta.created ?? Math.floor(Date.now() / 1000),
+      content: [content],
+      metadata: { ...DEFAULT_VISIBLE_MESSAGE_METADATA },
+    });
+  }
+
+  return messagesChange(state);
+}
+
+export function applyThoughtChunk(
+  state: AdapterState,
+  update: Extract<SessionNotification['update'], { sessionUpdate: 'agent_thought_chunk' }>
+): AcpChatStateChange[] {
+  if (update.content.type !== 'text') {
+    return [];
+  }
+
+  const gooseMeta = getGooseMessageMeta(update);
+  const messageId = update.messageId ?? gooseMeta.messageId;
+  const existing = findMessageForChunk(state, 'assistant', messageId, gooseMeta.created);
+
+  if (existing) {
+    const lastContent = existing.content[existing.content.length - 1];
+    if (lastContent?.type === 'thinking') {
+      lastContent.thinking += update.content.text;
+    } else {
+      existing.content.push({ type: 'thinking', thinking: update.content.text, signature: '' });
+    }
+  } else {
+    state.messages.push({
+      ...(messageId ? { id: messageId } : {}),
+      role: 'assistant',
+      created: gooseMeta.created ?? Math.floor(Date.now() / 1000),
+      content: [{ type: 'thinking', thinking: update.content.text, signature: '' }],
+      metadata: { ...DEFAULT_VISIBLE_MESSAGE_METADATA },
+    });
+  }
+
+  return messagesChange(state);
+}
+
+function messageContentFromAcpContentBlock(content: AcpContentBlock): MessageContent | undefined {
+  switch (content.type) {
+    case 'text':
+      return {
+        type: 'text',
+        text: content.text,
+        ...(content._meta ? { _meta: content._meta } : {}),
+        ...(content.annotations ? { annotations: content.annotations } : {}),
+      };
+    case 'image':
+      return {
+        type: 'image',
+        data: content.data,
+        mimeType: content.mimeType,
+        ...(content._meta ? { _meta: content._meta } : {}),
+        ...(content.annotations ? { annotations: content.annotations } : {}),
+      };
+    default:
+      return undefined;
+  }
+}
+
+export function findMessageForChunk(
+  state: AdapterState,
+  role: Message['role'],
+  messageId: string | undefined,
+  created: number | undefined
+): Message | undefined {
+  if (!messageId) {
+    return lastMergeableMessageWithRole(state, role);
+  }
+
+  const existing = state.messages.find(
+    (message) => message.id === messageId && message.role === role
+  );
+  if (existing) {
+    return existing;
+  }
+
+  const pending = lastMergeableMessageWithRole(state, role);
+  if (pending && !pending.id) {
+    pending.id = messageId;
+    pending.created = created ?? pending.created;
+    return pending;
+  }
+
+  return undefined;
+}
+
+function lastMergeableMessageWithRole(
+  state: AdapterState,
+  role: Message['role']
+): Message | undefined {
+  const lastMessage = state.messages[state.messages.length - 1];
+  if (lastMessage?.role !== role || lastMessage.metadata.agentVisible === false) {
+    return undefined;
+  }
+  return lastMessage;
+}
+
+function mergeTextChunk(existing: string, incoming: string): string {
+  if (!incoming || incoming === existing || existing.endsWith(incoming)) {
+    return existing;
+  }
+
+  if (!existing || incoming.startsWith(existing)) {
+    return incoming;
+  }
+
+  return existing + incoming;
+}
+
+function hasImageContent(message: Message, image: Extract<MessageContent, { type: 'image' }>) {
+  return message.content.some(
+    (content) =>
+      content.type === 'image' && content.data === image.data && content.mimeType === image.mimeType
+  );
+}
